Move symbolic state-related storage classes to new package symbolic.states.
