+++ prism/src/symbolic/comp/NondetModelChecker.java
@@ -73,10 +73,10 @@ import prism.PrismNative;
 import prism.PrismNotSupportedException;
 import prism.PrismSettings;
 import prism.PrismUtils;
-import prism.StateValues;
-import prism.StateValuesDV;
-import prism.StateValuesMTBDD;
-import prism.StateValuesVoid;
+import symbolic.states.StateValues;
+import symbolic.states.StateValuesDV;
+import symbolic.states.StateValuesMTBDD;
+import symbolic.states.StateValuesVoid;
 import prism.TileList;
 import sparse.PrismSparse;
 import strat.MDStrategyIV;
