+++ prism/src/symbolic/comp/ProbModelChecker.java
@@ -72,9 +72,9 @@ import prism.PrismNative;
 import prism.PrismNotSupportedException;
 import prism.PrismSettings;
 import prism.PrismUtils;
-import prism.StateValues;
-import prism.StateValuesDV;
-import prism.StateValuesMTBDD;
+import symbolic.states.StateValues;
+import symbolic.states.StateValuesDV;
+import symbolic.states.StateValuesMTBDD;
 import sparse.PrismSparse;
 import dv.DoubleVector;
 import explicit.ExportIterations;
