+++ prism/src/symbolic/comp/StochModelChecker.java
@@ -33,9 +33,9 @@ import dv.*;
 import mtbdd.*;
 import prism.Prism;
 import prism.PrismException;
-import prism.StateValues;
-import prism.StateValuesDV;
-import prism.StateValuesMTBDD;
+import symbolic.states.StateValues;
+import symbolic.states.StateValuesDV;
+import symbolic.states.StateValuesMTBDD;
 import sparse.*;
 import hybrid.*;
 import parser.ast.*;
