+++ prism/src/symbolic/model/ModelTransformation.java
@@ -28,8 +28,7 @@ package symbolic.model;
 
 import jdd.JDDNode;
 import prism.PrismException;
-import prism.StateValues;
-import symbolic.model.Model;
+import symbolic.states.StateValues;
 
 /**
  * Interface for a model transformation.
